+++ stringmetrics-py/src/lib.rs
@@ -3,7 +3,7 @@
 
 use pyo3::prelude::*;
 // use pyo3::types::PyIterator;
-use stringmetrics::algorithms;
+use ::stringmetrics::algorithms;
 
 /// Formats the sum of two numbers as string.
 #[pyfunction]